+++ cmd/hidden_lake/messenger/pkg/app/app.go
@@ -65,7 +65,6 @@ func (p *sApp) Run() error {
 	if p.fIsRun {
 		return errors.New("application already running")
 	}
-	p.fIsRun = true
 
 	if err := p.initDatabase(); err != nil {
 		return fmt.Errorf("open database: %w", err)
@@ -75,6 +74,7 @@ func (p *sApp) Run() error {
 	p.initInterfaceServiceHTTP()
 	p.initServicePPROF()
 
+	p.fIsRun = true
 	res := make(chan error)
 
 	go func() {
@@ -108,7 +108,7 @@ func (p *sApp) Run() error {
 	select {
 	case err := <-res:
 		resErr := fmt.Errorf("got run error: %w", err)
-		return utils.MergeErrors(resErr, p.Stop())
+		return utils.MergeErrors(resErr, p.stop())
 	case <-time.After(cInitStart):
 		p.fStdfLogger.PushInfo(fmt.Sprintf("%s is running...", pkg_settings.CServiceName))
 		return nil
@@ -119,6 +119,10 @@ func (p *sApp) Stop() error {
 	p.fMutex.Lock()
 	defer p.fMutex.Unlock()
 
+	return p.stop()
+}
+
+func (p *sApp) stop() error {
 	if !p.fIsRun {
 		return errors.New("application already stopped or not started")
 	}